YG-166 💡 #open #comment 관심사 분리 및 스타일 일부 변경
