+++ components/layouts/_components/search/type.ts
@@ -1,0 +1,8 @@
+export type modalContentProps = {
+    handleCloseSearch: () => void
+    debouncedSearchKeyword: (value: string) => void
+}
+
+export type modalTitleSectionProps = {
+    titleContainerStyle: string
+}
